+++ examples/kanban/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @olas/example-kanban
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @kontsedal/olas-core@0.0.3
+  - @kontsedal/olas-zod@0.0.3
+  - @kontsedal/olas-react@0.0.3
+  - @kontsedal/olas-persist@0.0.3
+  - @kontsedal/olas-devtools@0.0.3
+  - @kontsedal/olas-realtime@0.0.3
+  - @kontsedal/olas-cross-tab@0.0.3
+  - @kontsedal/olas-entities@0.0.3
+
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @kontsedal/olas-core@0.1.0
+  - @kontsedal/olas-zod@1.0.0
+  - @kontsedal/olas-react@1.0.0
+  - @kontsedal/olas-persist@1.0.0
+  - @kontsedal/olas-devtools@1.0.0
+  - @kontsedal/olas-realtime@1.0.0
+  - @kontsedal/olas-cross-tab@1.0.0
+  - @kontsedal/olas-entities@1.0.0
+
 ## 0.0.1
 
 ### Patch Changes